requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
bcrypt==4.1.2
Pillow==10.1.0